Move array_to_latex from latex.py to array.py
